fix test

+++ Dndc/TestDndcAst.c
@@ -202,7 +202,7 @@ TestFunction(TestParseAttributes){
 #include "dndc_node_types.h"
 PushDiagnostic();
 SuppressEnumCompare();
-#define X(a, b) _Static_assert(NODE_##a == DNDC_NODE_TYPE_##a, #a " doesn't match in public and private header!");
+#define X(a, b) _Static_assert((int)NODE_##a == (int)DNDC_NODE_TYPE_##a, #a " doesn't match in public and private header!");
 NODETYPES(X)
 #undef X
 PopDiagnostic();
